+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.3
+- Added new texture for the music discs (tanks to @YaCCBoy)
+- Removed unused code
+- Added support for option 'Notification Time'
+- Translation update
 # 1.2
 - Change to how musics.json is loaded. Ressource pack can now replace the default musics.json
 - Bumped Cloth config and Mod Menu version